[MOV] sale_by_packaging becomes sell_only_by_packaging
