--- sale_by_packaging/README.rst
@@ -1,117 +1,0 @@
-=================
-Sale By Packaging
-=================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Alpha-red.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Alpha
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--workflow-lightgray.png?logo=github
-    :target: https://github.com/OCA/sale-workflow/tree/14.0/sale_by_packaging
-    :alt: OCA/sale-workflow
-.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/sale-workflow-14-0/sale-workflow-14-0-sale_by_packaging
-    :alt: Translate me on Weblate
-.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/167/14.0
-    :alt: Try me on Runbot
-
-|badge1| |badge2| |badge3| |badge4| |badge5| 
-
-This module provides different configuration option to manage packagings on
-sale orders.
-
-The creation/update of sale order line will be blocked (by constraints) if the data on the
-sale.order.line does not fit with the configuration of the product's packagings.
-
-It's also possible to force the quantity to sell during creation/modification of the sale order line
-if the "Force sale quantity" is ticked on the packaging.
-
-For example, if your packaging is set to sell by 5 units and the employee fill
-the quantity with 3, the quantity will be automatically replaced by 5 (it always rounds up).
-
-.. IMPORTANT::
-   This is an alpha version, the data model and design can change at any time without warning.
-   Only for development or testing purpose, do not use in production.
-   `More details on development status <https://odoo-community.org/page/development-status>`_
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Configuration
-=============
-
-Following options are available to define which packaging type can be sold and
-which product can only be sold by packaging.
-
-* Can be sold: On product packaging type model, this checkbox defines if product
-  packagings from this particular type are available to be selected on sale
-  order line.
-
-* Sell only by packaging: On product template model, this checkbox restricts
-  sales of these products if no packaging is selected on the sale order line.
-  If no packaging is selected, it will either be auto-assigned if the quantity
-  on the sale order line matches a packaging quantity or an error will be raised.
-
-* Force sale quantity (on the packaging): force rounds up the quantity during
-  creation/modification of the sale order line with the factor set on the packaging.
-
-Known issues / Roadmap
-======================
-
-* Odoo allows to define product.packaging records with a qty = 0. This does not
-  make much sense with the can_be_sold checkbox, and we probably need to add a
-  constraint here.
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-workflow/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/sale-workflow/issues/new?body=module:%20sale_by_packaging%0Aversion:%2014.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* Camptocamp
-
-Contributors
-~~~~~~~~~~~~
-
-* Akim Juillerat <akim.juillerat@camptocamp.com>
-* Thomas Nowicki <thomas.nowicki@camptocamp.com>
-* François Honoré <francois.honore@acsone.eu>
-* Hiep (Nguyen Hoang) <hiepnh@trobz.com>
-* Phuc (Tran Thanh) <phuc@trobz.com>
-
-Maintainers
-~~~~~~~~~~~
-
-This module is maintained by the OCA.
-
-.. image:: https://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: https://odoo-community.org
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose
-mission is to support the collaborative development of Odoo features and
-promote its widespread use.
-
-This module is part of the `OCA/sale-workflow <https://github.com/OCA/sale-workflow/tree/14.0/sale_by_packaging>`_ project on GitHub.
-
-You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
