--- sale_by_packaging/__manifest__.py
@@ -1,22 +1,0 @@
-# Copyright 2020 Camptocamp SA
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl)
-{
-    "name": "Sale By Packaging",
-    "summary": "Manage sale of packaging",
-    "version": "16.0.1.0.0",
-    "development_status": "Alpha",
-    "category": "Warehouse Management",
-    "website": "https://github.com/OCA/sale-workflow",
-    "author": "Camptocamp, Odoo Community Association (OCA)",
-    "license": "AGPL-3",
-    "application": False,
-    "installable": True,
-    "depends": ["sale_order_line_packaging_qty", "product_packaging_level"],
-    "data": [
-        "views/product_packaging.xml",
-        "views/product_packaging_level.xml",
-        "views/product_template.xml",
-        "views/product_product.xml",
-        "views/sale_order_line.xml",
-    ],
-}
