--- sale_by_packaging/i18n/sale_by_packaging.pot
@@ -1,171 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* sale_by_packaging
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging__can_be_sold
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging_type__can_be_sold
-msgid "Can be sold"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,help:sale_by_packaging.field_product_packaging__force_sale_qty
-msgid ""
-"Determine if during the creation of a sale order line, the quantity should be forced with a multiple of the packaging.\n"
-"Example:\n"
-"You sell a product by packaging of 5 products.\n"
-"When the user will put 3 as quantity, the system can force the quantity to the superior unit (5 for this example)."
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging__display_name
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging_type__display_name
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_product__display_name
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_template__display_name
-#: model:ir.model.fields,field_description:sale_by_packaging.field_sale_order_line__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging__force_sale_qty
-msgid "Force sale quantity"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging__id
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging_type__id
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_product__id
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_template__id
-#: model:ir.model.fields,field_description:sale_by_packaging.field_sale_order_line__id
-msgid "ID"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging____last_update
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging_type____last_update
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_product____last_update
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_template____last_update
-#: model:ir.model.fields,field_description:sale_by_packaging.field_sale_order_line____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_product__min_sellable_qty
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_template__min_sellable_qty
-msgid "Min Sellable Qty"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,help:sale_by_packaging.field_product_product__min_sellable_qty
-#: model:ir.model.fields,help:sale_by_packaging.field_product_template__min_sellable_qty
-msgid ""
-"Minimum sellable quantity, according to the available packagings, if Only "
-"Sell by Packaging is set."
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_product__sell_only_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_template__sell_only_by_packaging
-msgid "Only sell by packaging"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,field_description:sale_by_packaging.field_product_packaging_type__packaging_ids
-msgid "Packaging"
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/sale_order_line.py:0
-#, python-format
-msgid ""
-"Packaging %s on product %s must be set as 'Can be sold' in order to be used "
-"on a sale order."
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/product_packaging_type.py:0
-#, python-format
-msgid ""
-"Packaging type \"{}\" must stay with \"Can be sold\", at least one product "
-"configured as \"sell only by packaging\" is using it."
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model,name:sale_by_packaging.model_product_product
-msgid "Product"
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/sale_order_line.py:0
-#, python-format
-msgid "Product %s can only be sold with a packaging and a packaging qantity."
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/product_template.py:0
-#, python-format
-msgid ""
-"Product %s cannot be defined to be sold only by packaging if it cannot be "
-"sold."
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/product_template.py:0
-#, python-format
-msgid ""
-"Product %s cannot be defined to be sold only by packaging if it does not "
-"have any packaging that can be sold defined."
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model,name:sale_by_packaging.model_product_packaging
-msgid "Product Packaging"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model,name:sale_by_packaging.model_product_template
-msgid "Product Template"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model.fields,help:sale_by_packaging.field_product_product__sell_only_by_packaging
-#: model:ir.model.fields,help:sale_by_packaging.field_product_template__sell_only_by_packaging
-msgid ""
-"Restrict the usage of this product on sale order lines without packaging "
-"defined"
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model,name:sale_by_packaging.model_sale_order_line
-msgid "Sales Order Line"
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/sale_order_line.py:0
-#, python-format
-msgid ""
-"This product packaging must be set as 'Can be sold' in order to be used on a"
-" sale order."
-msgstr ""
-
-#. module: sale_by_packaging
-#: model:ir.model,name:sale_by_packaging.model_product_packaging_type
-msgid "Type management for product.packaging"
-msgstr ""
-
-#. module: sale_by_packaging
-#: code:addons/sale_by_packaging/models/sale_order_line.py:0
-#, python-format
-msgid "Warning"
-msgstr ""
