--- sale_by_packaging/readme/CONTRIBUTORS.rst
@@ -1,5 +1,0 @@
-* Akim Juillerat <akim.juillerat@camptocamp.com>
-* Thomas Nowicki <thomas.nowicki@camptocamp.com>
-* François Honoré <francois.honore@acsone.eu>
-* Hiep (Nguyen Hoang) <hiepnh@trobz.com>
-* Phuc (Tran Thanh) <phuc@trobz.com>
