--- sale_by_packaging/readme/ROADMAP.rst
@@ -1,3 +1,0 @@
-* Odoo allows to define product.packaging records with a qty = 0. This does not
-  make much sense with the can_be_sold checkbox, and we probably need to add a
-  constraint here.
