--- sale_by_packaging/tests/__init__.py
@@ -1,4 +1,0 @@
-from . import test_minimum_sellable_qty
-from . import test_packaging_level_can_be_sold
-from . import test_sale_only_by_packaging
-from . import test_sale_line_onchanges
