--- sale_by_packaging/views/sale_order_line.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<odoo>
-    <record id="view_order_form_inherit_sale_stock_inherit" model="ir.ui.view">
-        <field name="name">sale.order.form.sale.stock.inherit</field>
-        <field name="model">sale.order</field>
-        <field name="inherit_id" ref="sale_stock.view_order_form_inherit_sale_stock" />
-        <field name="arch" type="xml">
-            <xpath
-                expr="//field[@name='order_line']//form//field[@name='product_packaging_id']"
-                position="attributes"
-            >
-                <attribute
-                    name="domain"
-                >[('product_id','=',product_id), ('can_be_sold', '=', True)]</attribute>
-            </xpath>
-            <xpath
-                expr="//field[@name='order_line']//tree//field[@name='product_packaging_id' and @optional='show']"
-                position="attributes"
-            >
-                <attribute
-                    name="domain"
-                >[('product_id','=',product_id), ('can_be_sold', '=', True)]</attribute>
-            </xpath>
-        </field>
-    </record>
-</odoo>
