+++ sell_only_by_packaging/i18n/sell_only_by_packaging.pot
@@ -1,0 +1,117 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* sell_only_by_packaging
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2023-07-27 09:54+0000\n"
+"PO-Revision-Date: 2023-07-27 09:54+0000\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: sell_only_by_packaging
+#: model:ir.model.fields,help:sell_only_by_packaging.field_product_packaging__force_sale_qty
+msgid ""
+"Determine if during the creation of a sale order line, the quantity should be forced with a multiple of the packaging.\n"
+"Example:\n"
+"You sell a product by packaging of 5 products.\n"
+"When the user will put 3 as quantity, the system can force the quantity to the superior unit (5 for this example)."
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model.fields,field_description:sell_only_by_packaging.field_product_packaging__force_sale_qty
+msgid "Force sale quantity"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model,name:sell_only_by_packaging.model_product_packaging_level
+msgid "Level management for product.packaging"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model.fields,field_description:sell_only_by_packaging.field_product_product__min_sellable_qty
+#: model:ir.model.fields,field_description:sell_only_by_packaging.field_product_template__min_sellable_qty
+msgid "Min Sellable Qty"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model.fields,help:sell_only_by_packaging.field_product_product__min_sellable_qty
+#: model:ir.model.fields,help:sell_only_by_packaging.field_product_template__min_sellable_qty
+msgid ""
+"Minimum sellable quantity, according to the available packagings, if Only "
+"Sell by Packaging is set."
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model.fields,field_description:sell_only_by_packaging.field_product_product__sell_only_by_packaging
+#: model:ir.model.fields,field_description:sell_only_by_packaging.field_product_template__sell_only_by_packaging
+msgid "Only sell by packaging"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#. odoo-python
+#: code:addons/sell_only_by_packaging/models/product_packaging_level.py:0
+#, python-format
+msgid ""
+"Packaging level \"{}\" must stay with \"Can be sold\", at least one product "
+"configured as \"sell only by packaging\" is using it."
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model,name:sell_only_by_packaging.model_product_template
+msgid "Product"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#. odoo-python
+#: code:addons/sell_only_by_packaging/models/sale_order_line.py:0
+#, python-format
+msgid "Product %s can only be sold with a packaging and a packaging qantity."
+msgstr ""
+
+#. module: sell_only_by_packaging
+#. odoo-python
+#: code:addons/sell_only_by_packaging/models/product_template.py:0
+#, python-format
+msgid ""
+"Product %s cannot be defined to be sold only by packaging if it cannot be "
+"sold."
+msgstr ""
+
+#. module: sell_only_by_packaging
+#. odoo-python
+#: code:addons/sell_only_by_packaging/models/product_template.py:0
+#, python-format
+msgid ""
+"Product %s cannot be defined to be sold only by packaging if it does not "
+"have any packaging that can be sold defined."
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model,name:sell_only_by_packaging.model_product_packaging
+msgid "Product Packaging"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model,name:sell_only_by_packaging.model_product_product
+msgid "Product Variant"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model.fields,help:sell_only_by_packaging.field_product_product__sell_only_by_packaging
+#: model:ir.model.fields,help:sell_only_by_packaging.field_product_template__sell_only_by_packaging
+msgid ""
+"Restrict the usage of this product on sale order lines without packaging "
+"defined"
+msgstr ""
+
+#. module: sell_only_by_packaging
+#: model:ir.model,name:sell_only_by_packaging.model_sale_order_line
+msgid "Sales Order Line"
+msgstr ""
