+++ sell_only_by_packaging/readme/CONTRIBUTORS.rst
@@ -1,0 +1,7 @@
+* Akim Juillerat <akim.juillerat@camptocamp.com>
+* Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
+* Thomas Nowicki <thomas.nowicki@camptocamp.com>
+* François Honoré <francois.honore@acsone.eu>
+* Hiep (Nguyen Hoang) <hiepnh@trobz.com>
+* Phuc (Tran Thanh) <phuc@trobz.com>
+* Duong (Tran Quoc) <duongtq@trobz.com>
