+++ sell_only_by_packaging/views/product_packaging.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+    <record id="product_packaging_tree_view_inherit" model="ir.ui.view">
+        <field name="name">product.packaging.tree.view.inherit</field>
+        <field name="model">product.packaging</field>
+        <field name="inherit_id" ref="product.product_packaging_tree_view" />
+        <field name="arch" type="xml">
+            <field name="qty" position="after">
+                <field name="force_sale_qty" />
+            </field>
+        </field>
+    </record>
+    <record id="product_packaging_form_view_inherit" model="ir.ui.view">
+        <field name="name">product.packaging.form.view.inherit</field>
+        <field name="model">product.packaging</field>
+        <field name="inherit_id" ref="product.product_packaging_form_view" />
+        <field name="arch" type="xml">
+            <field name="barcode" position="before">
+                <field name="force_sale_qty" />
+            </field>
+        </field>
+    </record>
+</odoo>
