+++ sell_only_by_packaging/views/product_template.xml
