--- setup/sale_by_packaging/odoo/addons/sale_by_packaging
@@ -1,1 +1,0 @@
-../../../../sale_by_packaging
