+++ setup/sell_only_by_packaging/odoo/addons/sell_only_by_packaging
@@ -1,0 +1,1 @@
+../../../../sell_only_by_packaging
